LIME-199: * Extract mocking data
