+++ frontend/src/bundles/overview/components/chart-goal-progress/chart-goal-progress.tsx
@@ -1,5 +1,3 @@
-import { type SingleValue } from 'react-select';
-
 import { Select } from '~/bundles/common/components/components.js';
 import { type SelectOption } from '~/bundles/common/components/select/types/types.js';
 import {
@@ -7,10 +5,13 @@ import {
     useCallback,
     useState,
 } from '~/bundles/common/hooks/hooks.js';
+import { type SingleValue } from '~/bundles/common/types/types.js';
 
 import { BarChart } from '../components.js';
+import { dataMappingSelectDatabase as dataMapping } from './helpers/data-select-database-mapping.helper.js';
+import { weeklyData } from './mock-database/mock-database-goal.js';
 
-const data: SelectOption[] = [
+const selectData: SelectOption[] = [
     {
         value: 'weekly',
         label: 'Weekly',
@@ -25,174 +26,10 @@ const data: SelectOption[] = [
     },
 ];
 
-const colors = {
-    workout: {
-        backgroundColor: 'rgba(224, 254, 16, 1)',
-    },
-    calories: {
-        backgroundColor: 'rgba(255, 94, 177, 1)',
-    },
-    steps: {
-        backgroundColor: 'rgba(126, 81, 255, 1)',
-    },
-};
-
-type BarChartData = {
-    labels: string[];
-    datasets: {
-        label: string;
-        data: number[];
-        backgroundColor: string;
-        borderColor: string;
-    }[];
-};
-
-const weeklyData = {
-    labels: [
-        'Sunday',
-        'Monday',
-        'Tuesday',
-        'Wednesday',
-        'Thursday',
-        'Friday',
-        'Saturday',
-    ],
-    datasets: [
-        {
-            label: 'Workout',
-            data: [20, 30, 40, 50, 60, 70, 80],
-            backgroundColor: colors.workout.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-        {
-            label: 'Calories',
-            data: [50, 60, 70, 80, 90, 100, 11],
-            backgroundColor: colors.calories.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-        {
-            label: 'Steps',
-            data: [100, 20, 30, 40, 50, 60, 70],
-            backgroundColor: colors.steps.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-    ],
-};
-
-const monthlyData = {
-    labels: ['Week 1', 'Week 2', 'Week 3', 'Week 4'],
-    datasets: [
-        {
-            label: 'Workout',
-            data: [20, 30, 40, 50],
-            backgroundColor: colors.workout.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-        {
-            label: 'Calories',
-            data: [50, 60, 70, 80],
-            backgroundColor: colors.calories.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-        {
-            label: 'Steps',
-            data: [100, 20, 30, 40],
-            backgroundColor: colors.steps.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-    ],
-};
-
-const yearlyData = {
-    labels: [
-        'January',
-        'February',
-        'March',
-        'April',
-        'May',
-        'June',
-        'July',
-        'August',
-        'September',
-        'October',
-        'November',
-        'December',
-    ],
-    datasets: [
-        {
-            label: 'Workout',
-            data: [20, 30, 40, 50, 60, 70, 80, 90, 100, 70, 50, 55],
-            backgroundColor: colors.workout.backgroundColor,
-            borderColor: 'transparent',
-            margin: 10,
-        },
-        {
-            label: 'Calories',
-            data: [50, 60, 70, 80, 90, 100, 100, 20, 30, 40, 50, 16],
-            backgroundColor: colors.calories.backgroundColor,
-
-            borderColor: 'transparent',
-        },
-        {
-            label: 'Steps',
-            data: [10, 20, 30, 40, 0, 60, 70, 8, 90, 100, 11, 12],
-            backgroundColor: colors.steps.backgroundColor,
-            borderColor: 'transparent',
-        },
-    ],
-};
-
 const ChartGoalProgress = (): JSX.Element => {
-    const [currentData, setCurrentData] = useState<SelectOption>({
-        value: 'weekly',
-        label: 'Weekly',
-    });
-
-    const optionsBarChart = {
-        scales: {
-            y: {
-                beginAtZero: true,
-                ticks: {
-                    min: 0,
-                    max: 100,
-                    stepSize: 20,
-                    callback: function (value: number): string {
-                        return `${value}%`;
-                    },
-                },
-            },
-        },
-        barThickness: 'flex',
-        borderRadius: 20,
-        barPercentage: 0.3,
-        responsive: true,
-        plugins: {
-            legend: {
-                display: true,
-                position: 'bottom',
-                align: 'start',
-                labels: {
-                    usePointStyle: true,
-                    pointStyle: 'circle',
-                    borderRadius: 20,
-                    padding: 20,
-                    boxWidth: 100,
-                },
-            },
-        },
-    };
-
-    const dataMapping: Record<string, BarChartData> = {
-        'weekly': weeklyData,
-        'monthly': monthlyData,
-        'yearly': yearlyData,
-    };
+    const [currentData, setCurrentData] = useState<SelectOption>(
+        selectData[0] as SelectOption,
+    );
 
     const { control, errors } = useAppForm({
         defaultValues: {
@@ -216,7 +53,7 @@ const ChartGoalProgress = (): JSX.Element => {
                         className="bg-secondary w-[100px] text-xs"
                         control={control}
                         name="select"
-                        options={data}
+                        options={selectData}
                         errors={errors}
                         value={currentData}
                         onChange={handleChange}
@@ -225,7 +62,6 @@ const ChartGoalProgress = (): JSX.Element => {
             </div>
             <BarChart
                 chartData={dataMapping[currentData.value] ?? weeklyData}
-                charOptions={optionsBarChart}
             />
         </div>
     );
