+++ frontend/src/bundles/overview/components/chart-goal-progress/helpers/data-select-database-mapping.helper.ts
@@ -1,0 +1,14 @@
+import { type BarChartData } from '../../charts/types/types.js';
+import {
+    monthlyData,
+    weeklyData,
+    yearlyData,
+} from '../mock-database/mock-database-goal.js';
+
+const dataMappingSelectDatabase: Record<string, BarChartData> = {
+    'weekly': weeklyData,
+    'monthly': monthlyData,
+    'yearly': yearlyData,
+};
+
+export { dataMappingSelectDatabase };
